refactor(store): Wave 4 pilot -- split the telemetry slice out of store.ts
First implementation-moving PR of Phase 2's Store split (the earlier
waves 0-3 were type-only). Proves the delegation mechanism the plan
settled on: each slice gets its own class per backend in
store/slices/<slice>.ts, and InMemoryStore/FirestoreStore become thin
composers holding one instance per slice plus one-line forwards.

telemetry was picked as the pilot because it's structurally the
cleanest slice in the interface: zero fields shared with any other
slice, zero cross-slice method calls, and -- unlike every other
slice -- no edge into deleteAccountIdentity at all. It needs none of
the query features fakeFirestore() doesn't support yet (orderBy,
ranges, in, cursors), so this PR proves the mechanism before that gap
has to be closed for the harder slices.

- New apps/api/src/store/slices/telemetry.ts: the TelemetryStore
  interface plus InMemoryTelemetryStore and FirestoreTelemetryStore,
  moved verbatim from store.ts (no logic changes -- confirmed by
  running the fake-vs-real reproduction the parity harness exists to
  catch, and the full test suite before/after).
- store.ts: both classes now hold a `telemetryStore` field and
  forward all four TelemetryStore methods to it. Store.telemetryStore
  (private field name only) and the TelemetryStore interface both
  still resolve through store.ts's existing exports, so none of the
  183 importers change. store.ts: 7447 -> 7369 lines.
- Dropped store.ts's re-export of the retention constants
  (TELEMETRY_TTL_FIELD, TELEMETRY_COLLECTION, VISIT_COLLECTION,
  telemetryExpiresAt, TELEMETRY_RETENTION_DAYS): their only consumers
  were the code that just moved, and store.test.ts, which now imports
  them directly from store/records/telemetry.js. The TelemetryEvent/
  VisitEvent/TelemetryEventType *type* re-export stays, since 9
  production files still import those types through store.js -- a
  deliberate scope boundary, not an oversight: migrating type-only
  imports across files unrelated to the slice split isn't required
  for this wave and would bloat a pilot meant to prove the mechanism
  cleanly.
- New apps/api/src/store/records/telemetry.test.ts: the
  `telemetryExpiresAt`/retention tests extracted verbatim from
  store.test.ts, since they test the pure record-module function, not
  InMemoryStore/FirestoreStore behaviour -- store.test.ts's own
  import list drops the three symbols this freed it from needing.

Comment-prose: both new files' baselines resealed scoped -- every
word moved is a verbatim relocation (store.ts -8092 words since Wave
1, store.test.ts -92 words matching telemetry.test.ts's +92 exactly),
none of it newly authored, so this isn't the loophole flagged on
PR #954's store-parity.test.ts.

Wave 4 plan: docs/phase2-store-split-execution-plan.md (private ops
repo). oauth is next -- the second-cleanest slice, and the one that
proves the small-documented-exception pattern for its
deleteAccountIdentity fragment.
