+++ eslint-rules/comment-prose-baseline.json
@@ -288,7 +288,9 @@
     "apps/api/src/store/records/review.ts": 488,
     "apps/api/src/store/records/social.ts": 232,
     "apps/api/src/store/records/submission.ts": 1562,
+    "apps/api/src/store/records/telemetry.test.ts": 92,
     "apps/api/src/store/records/telemetry.ts": 1136,
+    "apps/api/src/store/slices/telemetry.ts": 167,
     "apps/api/src/submission-status.test.ts": 35,
     "apps/api/src/submission-status.ts": 257,
     "apps/api/src/submissions.test.ts": 4563,
